add: split to_ts var_decl
