+++ src/to_ts_rust/to_ts/mod.rs
@@ -1,2 +1,3 @@
 pub mod to_ts;
 mod common_methods;
+mod variable_declaration;
