+++ src/to_ts_rust/to_ts/to_ts.rs
@@ -43,31 +43,7 @@ impl ToTs {
         }
     }
 
-    fn var_decl_to_ts(&mut self, var_decl: TypedVariableDeclaration) -> String {
-        let ident = var_decl.get_name();
-        let name = ident.get_name();
-        let type_name = var_decl.get_type_name();
-        let value = var_decl.get_value();
-
-        match &value {
-            Some(expr) => {
-                self.var_env.insert(ident, Some(CanAssignObj::TypedExpr(expr.clone())))
-            },
-            None => {self.var_env.insert(ident, None)},
-        };
-
-        match value {
-            Some(typed_expr) => {
-                match type_name {
-                    Some(type_flag) => format!("const {}:{}={};", name, self.type_flag_to_ts(type_flag), self.expr_to_ts(typed_expr)),
-                    None => format!("const {}={};", name, self.expr_to_ts(typed_expr)),
-                }
-            },
-            None => format!("let {};", name),
-        }
-    }
-
-    fn expr_to_ts(&self, typed_expr: TypedExpr) -> String {
+    pub fn expr_to_ts(&self, typed_expr: TypedExpr) -> String {
         match typed_expr {
             TypedExpr::NumExpr(_, num) => format!("{}", num.get_num()),
             TypedExpr::NumIdentExpr(_, ident) => {
@@ -206,49 +182,6 @@ mod tests {
     use crate::to_ts_rust::to_ts::to_ts::ToTs;
     use std::collections::HashMap;
     use crate::to_ts_rust::common_struct::CanAssignObj;
-
-    #[test]
-    fn test_no_type_var_declaration() {
-        let typed_stmts = vec![
-            TypedStmt::VariableDeclaration(
-                TypedVariableDeclaration::new(
-                    TypedIdent::new(
-                        "tmp1".to_string()),
-                    None,
-                    Some(
-                        TypedExpr::NumExpr(TypedAstType::Number, TypedNumber::new(0))
-                    ))
-                )
-        ];
-
-        let ts_code = ToTs::to_ts(typed_stmts, None);
-
-        let expected_ts_code = "const tmp1=0;";
-
-        assert_eq!(ts_code, expected_ts_code);
-
-    }
-
-    #[test]
-    fn test_has_type_var_declaration() {
-        let typed_stmts = vec![
-            TypedStmt::VariableDeclaration(
-                TypedVariableDeclaration::new(
-                    TypedIdent::new(
-                        "tmp1".to_string()),
-                    Some(TypeFlag::NumberType),
-                    Some(
-                        TypedExpr::NumExpr(TypedAstType::Number, TypedNumber::new(0))
-                    ))
-            )
-        ];
-
-        let ts_code = ToTs::to_ts(typed_stmts, None);
-
-        let expected_ts_code = "const tmp1:number=0;";
-
-        assert_eq!(ts_code, expected_ts_code);
-    }
 
     #[test]
     fn test_num_expr_stmt() {
